CHANGE: update phpdoc for local variables and minor changes
